requirements.txt
fastapi
uvicorn
psycopg2-binary
geoip2
python-multipart
requests